Feat: Renamed `FluxActions` to `FluxActionStack`
